Fixing a few things MSVC complained about. Also switched from *::value and *::type to *_v and *_t.

+++ src/gui/common.cpp
@@ -32,13 +32,13 @@ Size Size::em(size_t size)
 
 Size Size::em(double size)
 {
-	static const int em=text_height(is_large ? lfont_l : nfont);
-	return Size(size*em);
+	static const double em=text_height(is_large ? lfont_l : nfont);
+	return Size((int)(size*em));
 }
 
 Size Size::largePixels(int size)
 {
-	return Size(is_large ? size : (size*2.0/3.0));
+	return Size(is_large ? size : (size*2/3));
 }
 
 }
